Ввел систему крипов и атаку их по башням

+++ Fast Project/Assets/Scripts/Character/Stats.cs
@@ -6,7 +6,7 @@ public class Stats
 {
     public int Health;
     public int Damage;
-    public int AttackDistance;
+    public float AttackDistance;
     public float AttackCoolDown;
     public AnimationClip AttackAnimation, DeathAnimation;
 }
